feat: harmonize commands
